// frontend/src/types/types.ts
export interface IPost {
  id: string;
  privacy: "public" | "private" | "follower";
  photos: Record<string, any>[];
  comments: any[];
  description: string;
  likes: any[];
  author: IUser;
  commentsCount: number;
  likesCount: number;
  isBookmarked: boolean;
  isLiked: boolean;
  isOwnPost: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface IUser {
  id: string;
  username: string;
  fullName: string;
  email: string;
  profilePicture?: string | Record<string, any>;
  isEmailValidated: boolean;
  [prop: string]: any;
}

export interface IHelperState {
  targetComment: IComment | null;
  targetPost: IPost | null;
}

export interface IModalState {
  isOpenDeleteComment: boolean;
  isOpenDeletePost: boolean;
  isOpenEditPost: boolean;
  isOpenPostLikes: boolean;
}

export interface IComment {
  id: string;
  body: string;
  createdAt: Date;
  updatedAt: Date;
  isEdited: boolean;
  depth: number;
  replyCount: number;
  likesCount: number;
  post_id: string;
  isOwnComment: boolean;
  isPostOwner: boolean;
  isLiked: boolean;
  author: IUser;
}
// ...
export interface IPreferenceState {
  theme: "light" | "dark";
  language: "en" | "np" | "fr";
  targetComment: IComment | null;
  targetPost: IPost | null;
  hasSentVerificationMail: boolean;
  sendVerificationMailError: any;
  isOpenVerificationMessage: boolean;
}